+++ shaderforth/shader_program.cpp
@@ -298,9 +298,9 @@ void shader_program::run(image<uint32_t>& im)
       for (int i = 0; i < 8; ++i)
 #endif
         {
-        uint32_t red = (uint64_t)(clamp(r[i], 0.f, 1.f) * 255.f);
-        uint32_t green = (uint64_t)(clamp(g[i], 0.f, 1.f) * 255.f);
-        uint32_t blue = (uint64_t)(clamp(b[i], 0.f, 1.f) * 255.f);
+        uint32_t red = (uint32_t)(clamp(r[i], 0.f, 1.f) * 255.f);
+        uint32_t green = (uint32_t)(clamp(g[i], 0.f, 1.f) * 255.f);
+        uint32_t blue = (uint32_t)(clamp(b[i], 0.f, 1.f) * 255.f);
         *p_im++ = 0xff000000 | (blue << 16) | (green << 8) | red;
         }
       }
